requirements.txt
streamlit
langchain-community
langchain-core
faiss-cpu
langchainhub
tiktoken
langchain-ollama
